moved linksArray from Header.js to utils for more modularity
